Сделал рефакторинг кода и архитектуры
