+++ src/Screens/HomeScreen.js
@@ -1,0 +1,48 @@
+import React from "react";
+import { StyleSheet, View, ScrollView, ImageBackground } from "react-native";
+
+// CONFIG
+import colors from "../config/colors";
+
+import SearchComponent from "../components/HomeScreenComponents/SearchComponent";
+import CategoriesComponent from "../components/HomeScreenComponents/CategoriesComponent";
+import PlanetsComponent from "../components/HomeScreenComponents/PlanetsComponent";
+import HomepageAppbarComponent from "../components/HomeScreenComponents/HomepageAppbarComponent";
+
+function HomePageComponent() {
+  return (
+    <ImageBackground
+      style={styles.background}
+      source={require("../../assets/MainBackground.png")}
+    >
+      <View style={styles.container}>
+        <View style={styles.main}>
+          <HomepageAppbarComponent />
+          <SearchComponent />
+        </View>
+        <ScrollView>
+          <CategoriesComponent />
+          <PlanetsComponent />
+        </ScrollView>
+      </View>
+    </ImageBackground>
+  );
+}
+
+export default HomePageComponent;
+
+const styles = StyleSheet.create({
+  background: {
+    flex: 1,
+    resizeMode: "cover",
+    backgroundColor: colors.background,
+  },
+  container: {
+    fontFamily: "Roboto",
+    height: "100%",
+  },
+  main: {
+    paddingLeft: 20,
+    paddingRight: 20,
+  },
+});
