+++ src/components/HomeScreenComponents/PlanetComponent.js
@@ -2,8 +2,8 @@ import React from "react";
 import { StyleSheet, Text, View, TouchableOpacity, Image } from "react-native";
 
 // CONFIG
-import colors from "../config/colors";
-import ip from "../config/ip";
+import colors from "../../config/colors";
+import ip from "../../config/ip";
 
 // ICONS
 import { AntDesign } from "@expo/vector-icons";
@@ -11,19 +11,17 @@ import { AntDesign } from "@expo/vector-icons";
 export default function PlanetComponent({ planet, navigation }) {
   return (
     <TouchableOpacity onPress={() => navigation.navigate("Welcome")}>
-      <View style={styles.homepage__planets__block}>
+      <View style={styles.planet__block}>
         <Image
           source={{
             uri: `http://${ip}/${planet.planetImage}`,
             width: 110,
             height: 134,
           }}
-          style={styles.homepage__planets__block__image}
+          style={styles.planet__block__image}
         />
-        <View style={styles.homepage__planets__block__nav}>
-          <Text style={styles.homepage__planets__block__text}>
-            {planet.name}
-          </Text>
+        <View style={styles.planet__block__nav}>
+          <Text style={styles.planet__block__text}>{planet.name}</Text>
           <AntDesign name="arrowright" size={24} color="#EF5F53" />
         </View>
       </View>
@@ -32,7 +30,7 @@ export default function PlanetComponent({ planet, navigation }) {
 }
 
 const styles = StyleSheet.create({
-  homepage__planets__block: {
+  planet__block: {
     width: 140,
     height: 190,
     backgroundColor: colors.background,
@@ -45,19 +43,19 @@ const styles = StyleSheet.create({
     position: "relative",
     marginLeft: 16,
   },
-  homepage__planets__block__image: {
+  planet__block__image: {
     position: "absolute",
     top: 0,
     left: 0,
     borderTopLeftRadius: 8,
   },
-  homepage__planets__block__nav: {
+  planet__block__nav: {
     flexDirection: "row",
     width: "100%",
     justifyContent: "space-between",
     alignItems: "center",
   },
-  homepage__planets__block__text: {
+  planet__block__text: {
     color: colors.white,
     fontSize: 16,
     fontWeight: "bold",
